Refactor codebase to modular structure with improved configuration management
- Introduced a new `app` package to encapsulate application logic.
- Created `config.py` for centralized configuration management using environment variables.
- Implemented `http.py` for HTTP session management with retry logic.
- Developed `main.py` to handle application startup and scheduling using cron expressions.
- Added `models.py` for data structures, specifically `UpdateInfo`.
- Created `notifications` package for handling webhook notifications.
- Established `registry` package for managing Docker image registries (DockerHub and GHCR).
- Refactored `scanner.py` to utilize new modular structure for checking updates on Docker containers.
- Updated tests to reflect changes in module imports and structure.
- Ensured backward compatibility with existing functionality while enhancing code organization.
