+++ tests/test_registry.py
@@ -3,36 +3,39 @@
 from unittest.mock import MagicMock, patch
 import requests
 
-import monitor
+from app import http as http_mod
+from app.registry import fetch_all_tags
+from app.registry.dockerhub import get_dockerhub_token, _fetch_dockerhub_tags
+from app.registry.ghcr import _get_ghcr_token, _fetch_ghcr_tags
 
 
 class TestGetDockerhubToken:
     """Tests for get_dockerhub_token()."""
 
-    @patch.object(monitor, "http_session")
+    @patch.object(http_mod, "http_session")
     def test_no_credentials_returns_none(self, mock_session):
-        assert monitor.get_dockerhub_token("", "") is None
+        assert get_dockerhub_token("", "") is None
         mock_session.post.assert_not_called()
 
-    @patch.object(monitor, "http_session")
+    @patch.object(http_mod, "http_session")
     def test_no_username_returns_none(self, mock_session):
-        assert monitor.get_dockerhub_token("", "password") is None
+        assert get_dockerhub_token("", "password") is None
 
-    @patch.object(monitor, "http_session")
+    @patch.object(http_mod, "http_session")
     def test_no_password_returns_none(self, mock_session):
-        assert monitor.get_dockerhub_token("user", "") is None
+        assert get_dockerhub_token("user", "") is None
 
-    @patch.object(monitor, "http_session")
+    @patch.object(http_mod, "http_session")
     def test_auth_failure_returns_none(self, mock_session):
         mock_session.post.side_effect = Exception("401 Unauthorized")
-        result = monitor.get_dockerhub_token("user", "badpass")
+        result = get_dockerhub_token("user", "badpass")
         assert result is None
 
 
 class TestFetchDockerhubTags:
     """Tests for _fetch_dockerhub_tags()."""
 
-    @patch.object(monitor, "http_session")
+    @patch.object(http_mod, "http_session")
     def test_single_page_returns_tags(self, mock_session):
         mock_resp = MagicMock()
         mock_resp.json.return_value = {
@@ -42,10 +45,10 @@ class TestFetchDockerhubTags:
         mock_resp.raise_for_status = MagicMock()
         mock_session.get.return_value = mock_resp
 
-        tags = monitor._fetch_dockerhub_tags("nginx", "token")
+        tags = _fetch_dockerhub_tags("nginx", "token")
         assert tags == ["1.0.0", "1.1.0"]
 
-    @patch.object(monitor, "http_session")
+    @patch.object(http_mod, "http_session")
     def test_early_stop_on_current_tag(self, mock_session):
         mock_resp = MagicMock()
         mock_resp.json.return_value = {
@@ -55,77 +58,77 @@ class TestFetchDockerhubTags:
         mock_resp.raise_for_status = MagicMock()
         mock_session.get.return_value = mock_resp
 
-        tags = monitor._fetch_dockerhub_tags("nginx", "token", current_tag="1.0.0")
+        tags = _fetch_dockerhub_tags("nginx", "token", current_tag="1.0.0")
         # Should stop after first page since current_tag was found
         assert "1.0.0" in tags
         assert mock_session.get.call_count == 1
 
-    @patch.object(monitor, "http_session")
+    @patch.object(http_mod, "http_session")
     def test_strips_docker_io_prefix(self, mock_session):
         mock_resp = MagicMock()
         mock_resp.json.return_value = {"results": [{"name": "latest"}], "next": None}
         mock_resp.raise_for_status = MagicMock()
         mock_session.get.return_value = mock_resp
 
-        monitor._fetch_dockerhub_tags("docker.io/library/nginx", "token")
+        _fetch_dockerhub_tags("docker.io/library/nginx", "token")
         url = mock_session.get.call_args[0][0]
         assert "library" in url and "nginx" in url
 
-    @patch.object(monitor, "http_session")
+    @patch.object(http_mod, "http_session")
     def test_http_error_returns_partial(self, mock_session):
         mock_session.get.side_effect = requests.HTTPError("500 Server Error")
-        tags = monitor._fetch_dockerhub_tags("nginx", "token")
+        tags = _fetch_dockerhub_tags("nginx", "token")
         assert tags == []
 
-    @patch.object(monitor, "http_session")
+    @patch.object(http_mod, "http_session")
     def test_general_exception_returns_empty(self, mock_session):
         mock_session.get.side_effect = Exception("network timeout")
-        tags = monitor._fetch_dockerhub_tags("nginx", "token")
+        tags = _fetch_dockerhub_tags("nginx", "token")
         assert tags == []
 
 
 class TestGetGhcrToken:
     """Tests for _get_ghcr_token()."""
 
-    @patch.object(monitor, "http_session")
+    @patch.object(http_mod, "http_session")
     def test_successful_token_exchange(self, mock_session):
         mock_resp = MagicMock()
         mock_resp.json.return_value = {"token": "ghcr-token-abc"}
         mock_resp.raise_for_status = MagicMock()
         mock_session.get.return_value = mock_resp
 
-        result = monitor._get_ghcr_token("owner", "repo", "gh_pat_xxx")
+        result = _get_ghcr_token("owner", "repo", "gh_pat_xxx")
         assert result == "ghcr-token-abc"
 
-    @patch.object(monitor, "http_session")
+    @patch.object(http_mod, "http_session")
     def test_failed_token_exchange_returns_none(self, mock_session):
         mock_session.get.side_effect = Exception("403 Forbidden")
-        result = monitor._get_ghcr_token("owner", "repo", "gh_pat_xxx")
+        result = _get_ghcr_token("owner", "repo", "gh_pat_xxx")
         assert result is None
 
 
 class TestFetchGhcrTags:
     """Tests for _fetch_ghcr_tags()."""
 
-    @patch.object(monitor, "http_session")
+    @patch.object(http_mod, "http_session")
     def test_no_github_token_returns_empty(self, mock_session):
-        tags = monitor._fetch_ghcr_tags("ghcr.io/owner/repo", "")
+        tags = _fetch_ghcr_tags("ghcr.io/owner/repo", "")
         assert tags == []
         mock_session.get.assert_not_called()
 
-    @patch.object(monitor, "http_session")
+    @patch.object(http_mod, "http_session")
     def test_invalid_image_path_returns_empty(self, mock_session):
-        tags = monitor._fetch_ghcr_tags("ghcr.io/invalid", "gh_token")
+        tags = _fetch_ghcr_tags("ghcr.io/invalid", "gh_token")
         assert tags == []
 
-    @patch("monitor._get_ghcr_token", return_value=None)
-    @patch.object(monitor, "http_session")
+    @patch("app.registry.ghcr._get_ghcr_token", return_value=None)
+    @patch.object(http_mod, "http_session")
     def test_no_pull_token_returns_empty(self, mock_session, mock_get_token):
-        tags = monitor._fetch_ghcr_tags("ghcr.io/owner/repo", "gh_token")
+        tags = _fetch_ghcr_tags("ghcr.io/owner/repo", "gh_token")
         assert tags == []
 
-    @patch("monitor._get_ghcr_token", return_value="pull-token")
-    @patch.object(monitor, "http_session")
+    @patch("app.registry.ghcr._get_ghcr_token", return_value="pull-token")
+    @patch.object(http_mod, "http_session")
     def test_successful_fetch(self, mock_session, mock_get_token):
         mock_resp = MagicMock()
         mock_resp.json.return_value = {"tags": ["v1.0.0", "v1.1.0"]}
@@ -133,11 +136,11 @@ class TestFetchGhcrTags:
         mock_resp.headers = {"Link": ""}
         mock_session.get.return_value = mock_resp
 
-        tags = monitor._fetch_ghcr_tags("ghcr.io/owner/repo", "gh_token")
+        tags = _fetch_ghcr_tags("ghcr.io/owner/repo", "gh_token")
         assert tags == ["v1.0.0", "v1.1.0"]
 
-    @patch("monitor._get_ghcr_token", return_value="pull-token")
-    @patch.object(monitor, "http_session")
+    @patch("app.registry.ghcr._get_ghcr_token", return_value="pull-token")
+    @patch.object(http_mod, "http_session")
     def test_pagination_with_link_header(self, mock_session, mock_get_token):
         resp1 = MagicMock()
         resp1.json.return_value = {"tags": ["v1.0.0"]}
@@ -151,40 +154,40 @@ class TestFetchGhcrTags:
 
         mock_session.get.side_effect = [resp1, resp2]
 
-        tags = monitor._fetch_ghcr_tags("ghcr.io/owner/repo", "gh_token")
+        tags = _fetch_ghcr_tags("ghcr.io/owner/repo", "gh_token")
         assert tags == ["v1.0.0", "v2.0.0"]
         assert mock_session.get.call_count == 2
 
-    @patch("monitor._get_ghcr_token", return_value="pull-token")
-    @patch.object(monitor, "http_session")
+    @patch("app.registry.ghcr._get_ghcr_token", return_value="pull-token")
+    @patch.object(http_mod, "http_session")
     def test_http_error_returns_partial(self, mock_session, mock_get_token):
         mock_session.get.side_effect = requests.HTTPError("403 Forbidden")
-        tags = monitor._fetch_ghcr_tags("ghcr.io/owner/repo", "gh_token")
+        tags = _fetch_ghcr_tags("ghcr.io/owner/repo", "gh_token")
         assert tags == []
 
-    @patch("monitor._get_ghcr_token", return_value="pull-token")
-    @patch.object(monitor, "http_session")
+    @patch("app.registry.ghcr._get_ghcr_token", return_value="pull-token")
+    @patch.object(http_mod, "http_session")
     def test_general_exception_returns_empty(self, mock_session, mock_get_token):
         mock_session.get.side_effect = Exception("connection reset")
-        tags = monitor._fetch_ghcr_tags("ghcr.io/owner/repo", "gh_token")
+        tags = _fetch_ghcr_tags("ghcr.io/owner/repo", "gh_token")
         assert tags == []
 
 
 class TestFetchAllTags:
     """Tests for fetch_all_tags() routing."""
 
-    @patch("monitor._fetch_dockerhub_tags", return_value=["1.0.0"])
+    @patch("app.registry._fetch_dockerhub_tags", return_value=["1.0.0"])
     def test_routes_dockerhub(self, mock_fetch):
-        tags = monitor.fetch_all_tags("nginx", "token", "")
+        tags = fetch_all_tags("nginx", "token", "")
         assert tags == ["1.0.0"]
         mock_fetch.assert_called_once_with("nginx", "token", None)
 
-    @patch("monitor._fetch_ghcr_tags", return_value=["v1.0.0"])
+    @patch("app.registry._fetch_ghcr_tags", return_value=["v1.0.0"])
     def test_routes_ghcr(self, mock_fetch):
-        tags = monitor.fetch_all_tags("ghcr.io/owner/repo", None, "gh_token")
+        tags = fetch_all_tags("ghcr.io/owner/repo", None, "gh_token")
         assert tags == ["v1.0.0"]
         mock_fetch.assert_called_once_with("ghcr.io/owner/repo", "gh_token")
 
     def test_unknown_registry_returns_empty(self):
-        tags = monitor.fetch_all_tags("quay.io/org/image", None, "")
+        tags = fetch_all_tags("quay.io/org/image", None, "")
         assert tags == []
